made complex rest node relationships operation smarter

+++ src/main/java/me/phifty/graph/neo4j/Neo4jComplex.java
@@ -1,13 +1,11 @@
 package me.phifty.graph.neo4j;
 
 import me.phifty.graph.Complex;
+import me.phifty.graph.ComplexResetNodeRelationshipsResult;
 import me.phifty.graph.Handler;
 import org.neo4j.graphdb.*;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 /**
  * @author phifty <b.phifty@gmail.com>
@@ -60,35 +58,51 @@ public class Neo4jComplex implements Complex {
   }
 
   @Override
-  public void resetNodeRelationships(Object nodeId, String name, Iterable<Object> targetIds, Handler<Iterable<Object>> handler) {
-    List<Object> notFoundNodeIds = new ArrayList<>();
+  public void resetNodeRelationships(Object nodeId, String name, Set<Object> targetIds, Handler<ComplexResetNodeRelationshipsResult> handler) {
     Transaction transaction = graphDatabaseService.beginTx();
     try {
       Node node = finder.getNode(nodeId);
+      RelationshipType relationshipType = DynamicRelationshipType.forName(name);
 
-      isolateNode(node);
+      Set<Object> addedNodeIds = new HashSet<>();
+      Set<Object> removedNodeIds = new HashSet<>();
+      Set<Object> notFoundNodeIds = new HashSet<>();
+
+      Iterable<Relationship> relationships = node.getRelationships(relationshipType);
+
+      // remove obsolete nodes
+      Set<Object> currentlyRelatedNodeIds = new HashSet<>();
+      for (Relationship relationship : relationships) {
+        Node targetNode = relationship.getOtherNode(node);
+        Object targetNodeId = finder.getNodeId(targetNode);
+        currentlyRelatedNodeIds.add(targetNodeId);
+
+        if (!targetIds.contains(targetNodeId)) {
+          relationship.delete();
+          removedNodeIds.add(targetNodeId);
+        }
+      }
+
+      // add missing nodes
       for (Object targetId : targetIds) {
-        Node targetNode = finder.getNode(targetId);
-        if (targetNode == null) {
-          notFoundNodeIds.add(targetId);
-        } else {
-          node.createRelationshipTo(targetNode, DynamicRelationshipType.forName(name));
+        if (!currentlyRelatedNodeIds.contains(targetId)) {
+          Node targetNode = finder.getNode(targetId);
+          if (targetNode == null) {
+            notFoundNodeIds.add(targetId);
+          } else {
+            node.createRelationshipTo(targetNode, relationshipType);
+            addedNodeIds.add(targetId);
+          }
         }
       }
 
       transaction.success();
-      handler.handle(notFoundNodeIds);
+      handler.handle(new ComplexResetNodeRelationshipsResult(addedNodeIds, removedNodeIds, notFoundNodeIds));
     } catch (Exception exception) {
       transaction.failure();
       handler.exception(exception);
     } finally {
       transaction.finish();
-    }
-  }
-
-  private void isolateNode(Node node) {
-    for (Relationship relationship : node.getRelationships()) {
-      relationship.delete();
     }
   }
